Add meta diagnostics router and startup logging
